requirements.txt
prefect==3.0.0
pandas
requests
aiohttp
aiofiles
aiogram
python-dotenv
